Restructre graphql directory
