+++ server/graphql/resolvers.js
@@ -1,0 +1,15 @@
+
+const books = [
+  {
+    title: "Harry Potter and the Sorcerer's stone",
+    author: 'J.K. Rowling',
+  },
+  {
+    title: 'Jurassic Park',
+    author: 'Michael Crichton',
+  },
+];
+
+export default {
+  Query: { books: () => books }
+};
