--- server/graphql/schema.js
@@ -1,28 +1,0 @@
-import {makeExecutableSchema} from "graphql-tools/";
-
-const books = [
-  {
-    title: "Harry Potter and the Sorcerer's stone",
-    author: 'J.K. Rowling',
-  },
-  {
-    title: 'Jurassic Park',
-    author: 'Michael Crichton',
-  },
-];
-
-const typeDefs = `
-  type Query { books: [Book] }
-  type Book { title: String, author: String }
-`;
-
-const resolvers = {
-  Query: { books: () => books }
-};
-
-const schema = makeExecutableSchema({
-  typeDefs,
-  resolvers
-});
-
-export default schema;
