+++ server/graphql/typeDefs.js
@@ -1,0 +1,4 @@
+export default `
+  type Query { books: [Book] }
+  type Book { title: String, author: String }
+`;
