+++ server/middlewares/graphql.js
@@ -1,7 +1,14 @@
-import schema from "../graphql/schema";
 import {graphiqlExpress, graphqlExpress} from "apollo-server-express";
+import {makeExecutableSchema} from "graphql-tools/dist/index";
+import resolvers from "../graphql/resolvers";
+import typeDefs from "../graphql/typeDefs";
 
-const graphql = graphqlExpress({schema});
+const schema = makeExecutableSchema({
+  typeDefs,
+  resolvers
+});
+
+const graphql = graphqlExpress({ schema });
 const graphiql = graphiqlExpress({ endpointURL: '/graphql' });
 
 export { graphql, graphiql };
